+++ spec/audited_spec_helpers.rb
@@ -8,8 +8,8 @@ module AuditedSpecHelpers
     Models::ActiveRecord::User.new({name: 'darth', username: 'darth', password: 'noooooooo'}.merge(attrs))
   end
 
-  def create_versions(n = 2)
-    Models::ActiveRecord::User.create(name: 'Foobar 1').tap do |u|
+  def create_versions(n = 2, attrs = {})
+    Models::ActiveRecord::User.create(name: 'Foobar 1', **attrs).tap do |u|
       (n - 1).times do |i|
         u.update_attribute :name, "Foobar #{i + 2}"
       end